+++ Changelog.md
@@ -1,5 +1,10 @@
+## 0.2.3
+- Fixes to infinite shadow polygons
+- Add `invertKey` and `fromAngle` methods.
+
 ## 0.2.2
 - Don't use cached elevation properties, to avoid caching problems and simplify approach.
+- Use token topE for VisionSources.
 
 ## 0.2.1
 - Fix for PIXI.Point.key getter.